Example usage of VERY BASIC Anti CSRF Token
Not tested - only PoC ;p but should work

+++ actions/removeBan.php
@@ -1,13 +1,18 @@
 <?php
 include('../template-parts/session.php');
 require '../config.php';
+require '../functions.php';
+
 $banid = $_GET['banid'];
 
+if(isValidCsrfToken($_GET['csrfToken'])) {
+	$sql = "DELETE FROM ".USERS_BANNED_TABLE." WHERE ".USERS_BANNED_ID_COLUMN." = '{$banid}' ";
 
-$sql = "DELETE FROM ".USERS_BANNED_TABLE." WHERE ".USERS_BANNED_ID_COLUMN." = '{$banid}' ";
-
-if ($link->query($sql) === TRUE) {
-    echo "Ban removed";
+	if ($link->query($sql) === TRUE) {
+		echo "Ban removed";
+	} else {
+		echo "Error: " . $sql . "<br>" . $conn->error;
+	}
 } else {
-    echo "Error: " . $sql . "<br>" . $conn->error;
+	echo "Invalid Anti CSRF Token";
 }
